3/11 commit

+++ Ps 1일 n문제/210310/2225/2225.cpp
@@ -1,7 +1,7 @@
 #include<iostream>
 using namespace std;
 
-long long arr[201][201];
+int arr[201][201];
 
 int main(){
 	int n,k;
